refactor: move validation deps to own module
